# haven/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Haven API
  version: 0.1.0
  description: |
    Haven is an AI-powered housing marketplace matching seekers and landlords
    based on lifestyle compatibility, verified trust scores, and real-time chat.
  contact:
    name: Haven Engineering
    email: engineering@haven.app
servers:
  - url: https://haven.app
    description: Production
  - url: https://staging.haven.app
    description: Staging
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT access token
  schemas:
...
    Match:
      type: object
      properties:
        id:
          type: string
          format: uuid
        listing_id:
          type: string
          format: uuid
        total_score:
          type: integer
          minimum: 0
          maximum: 100
        lifestyle_score:
          type: integer
        personality_score:
          type: integer
        location_score:
          type: integer
        budget_score:
          type: integer
        amenity_score:
          type: integer
        trust_score:
          type: integer
...
paths:
...
  /api/matches/score:
    post:
      operationId: scoreMatch
      summary: Compute a match score between a seeker and a listing
      tags: [Matching]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [seeker_id, listing_id]
              properties:
                seeker_id:
                  type: string
                  format: uuid
                listing_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Match score result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
...
  /api/ai/suggest-pricing:
    post:
      operationId: suggestPricing
      summary: AI pricing suggestion based on market comparables
      tags: [AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [city, state, bedrooms, sqft]
              properties:
                city:
                  type: string
                state:
                  type: string
                bedrooms:
                  type: integer
                sqft:
                  type: integer
      responses:
        '200':
          description: Pricing suggestion
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggested_price:
                    type: integer
                  range_low:
                    type: integer
                  range_high:
                    type: integer
                  confidence:
                    type: number
...
  /api/payments/webhook:
    post:
      operationId: stripeWebhook
      summary: Stripe webhook receiver
      description: |
        Receives signed events from Stripe. Must include a valid
        `Stripe-Signature` header. This endpoint is called by Stripe only and
        should not be invoked by application clients.
      security: []
      tags: [Payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Raw Stripe Event object
      responses:
        '200':
          description: Event received and processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                  skipped:
                    type: string
                    description: Present when event is a duplicate
        '400':
          description: Missing or invalid Stripe-Signature
  /api/payments/refund:
    post:
      operationId: createRefund
      summary: Issue a refund for a booking
      tags: [Payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId]
              properties:
                bookingId:
                  type: string
                  format: uuid
                reason:
                  type: string
                  enum: [duplicate, fraudulent, requested_by_customer]
      responses:
        '200':
          description: Refund created
          content:
            application/json:
              schema:
                type: object
                properties:
                  refundId:
                    type: string
                  status:
                    type: string
        '403':
          description: Forbidden — not a booking participant
        '404':
          description: Booking not found
        '409':
          description: Already refunded
        '422':
          description: No payment found for this booking
...
  /api/ai/analyze-photos:
    post:
      operationId: analyzePhotos
      summary: AI analysis of listing photos
      tags: [AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [imageUrls]
              properties:
                imageUrls:
                  type: array
                  items:
                    type: string
                    format: uri
                  minItems: 1
                  maxItems: 20
      responses:
        '200':
          description: Photo analysis results
        '400':
          description: Validation error (no images or too many)
...
  /api/ai/voice-to-listing:
    post:
      operationId: voiceToListing
      summary: Convert audio description to listing draft
      tags: [AI]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
                  description: Audio file (max 25 MB, formats supported by OpenAI Whisper)
      responses:
        '200':
          description: Transcription and generated listing draft
        '400':
          description: No audio file provided
        '413':
          description: Audio file too large (max 25 MB)
...